Updated version of the Monte Carlo simulations
I have modified the code in such a way that, we can simulate the 3D vector position of rotation angles (using 2 angles). Then the angles are projected in the sky.

What is currently missing is 1) checking whether normal distribution for the added uncertainties are a good approximation. 2) Maybe try linear or non-uniform distribution when selecting the samples. This can be consulted with Hsi-wei